+++ tools/node-depman/tests/download_test.rs
@@ -1,4 +1,5 @@
 use proto_pdk_test_utils::*;
+use rustc_hash::FxHashMap;
 
 #[derive(Debug, Default, serde::Deserialize, serde::Serialize)]
 #[serde(default, deny_unknown_fields, rename_all = "kebab-case")]
@@ -98,6 +99,76 @@ mod node_depman_tool {
                 }
             );
         }
+
+        #[tokio::test(flavor = "multi_thread")]
+        async fn extracts_auth_token_header() {
+            let sandbox = create_empty_proto_sandbox();
+            sandbox.create_file(".npmrc", "//registry.npmjs.org/:_authToken = abc123");
+
+            let plugin = sandbox
+                .create_plugin_with_config("npm-test", |config| {
+                    config.host(HostOS::Linux, HostArch::Arm64);
+                })
+                .await;
+
+            assert_eq!(
+                plugin
+                    .download_prebuilt(DownloadPrebuiltInput {
+                        context: PluginContext {
+                            version: VersionSpec::parse("9.0.0").unwrap(),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    })
+                    .await,
+                DownloadPrebuiltOutput {
+                    archive_prefix: Some("package".into()),
+                    download_url: "https://registry.npmjs.org/npm/-/npm-9.0.0.tgz".into(),
+                    http_headers: FxHashMap::from_iter([(
+                        "Authorization".into(),
+                        "Bearer abc123".into()
+                    )]),
+                    ..Default::default()
+                }
+            );
+        }
+
+        #[tokio::test(flavor = "multi_thread")]
+        async fn extracts_auth_token_header_for_custom_registry() {
+            let sandbox = create_empty_proto_sandbox();
+            sandbox.create_file(".npmrc", "//registry.yarnpkg.com/:_authToken = abc123");
+
+            let plugin = sandbox
+                .create_plugin_with_config("npm-test", |config| {
+                    config.host(HostOS::Linux, HostArch::Arm64).tool_config(
+                        NodeDepmanPluginConfig {
+                            registry_url: "https://registry.yarnpkg.com".into(),
+                        },
+                    );
+                })
+                .await;
+
+            assert_eq!(
+                plugin
+                    .download_prebuilt(DownloadPrebuiltInput {
+                        context: PluginContext {
+                            version: VersionSpec::parse("9.0.0").unwrap(),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    })
+                    .await,
+                DownloadPrebuiltOutput {
+                    archive_prefix: Some("package".into()),
+                    download_url: "https://registry.yarnpkg.com/npm/-/npm-9.0.0.tgz".into(),
+                    http_headers: FxHashMap::from_iter([(
+                        "Authorization".into(),
+                        "Bearer abc123".into()
+                    )]),
+                    ..Default::default()
+                }
+            );
+        }
     }
 
     mod pnpm {
@@ -185,6 +256,76 @@ mod node_depman_tool {
                 DownloadPrebuiltOutput {
                     archive_prefix: Some("package".into()),
                     download_url: "https://some-internal-url.example/pnpm/-/pnpm-8.0.0.tgz".into(),
+                    ..Default::default()
+                }
+            );
+        }
+
+        #[tokio::test(flavor = "multi_thread")]
+        async fn extracts_auth_token_header() {
+            let sandbox = create_empty_proto_sandbox();
+            sandbox.create_file(".npmrc", "//registry.npmjs.org/:_authToken = abc123");
+
+            let plugin = sandbox
+                .create_plugin_with_config("pnpm-test", |config| {
+                    config.host(HostOS::Linux, HostArch::Arm64);
+                })
+                .await;
+
+            assert_eq!(
+                plugin
+                    .download_prebuilt(DownloadPrebuiltInput {
+                        context: PluginContext {
+                            version: VersionSpec::parse("9.0.0").unwrap(),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    })
+                    .await,
+                DownloadPrebuiltOutput {
+                    archive_prefix: Some("package".into()),
+                    download_url: "https://registry.npmjs.org/npm/-/npm-9.0.0.tgz".into(),
+                    http_headers: FxHashMap::from_iter([(
+                        "Authorization".into(),
+                        "Bearer abc123".into()
+                    )]),
+                    ..Default::default()
+                }
+            );
+        }
+
+        #[tokio::test(flavor = "multi_thread")]
+        async fn extracts_auth_token_header_for_custom_registry() {
+            let sandbox = create_empty_proto_sandbox();
+            sandbox.create_file(".npmrc", "//registry.yarnpkg.com/:_authToken = abc123");
+
+            let plugin = sandbox
+                .create_plugin_with_config("pnpm-test", |config| {
+                    config.host(HostOS::Linux, HostArch::Arm64).tool_config(
+                        NodeDepmanPluginConfig {
+                            registry_url: "https://registry.yarnpkg.com".into(),
+                        },
+                    );
+                })
+                .await;
+
+            assert_eq!(
+                plugin
+                    .download_prebuilt(DownloadPrebuiltInput {
+                        context: PluginContext {
+                            version: VersionSpec::parse("9.0.0").unwrap(),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    })
+                    .await,
+                DownloadPrebuiltOutput {
+                    archive_prefix: Some("package".into()),
+                    download_url: "https://registry.yarnpkg.com/npm/-/npm-9.0.0.tgz".into(),
+                    http_headers: FxHashMap::from_iter([(
+                        "Authorization".into(),
+                        "Bearer abc123".into()
+                    )]),
                     ..Default::default()
                 }
             );
@@ -346,6 +487,85 @@ mod node_depman_tool {
                     .unwrap()
                     .exe_path,
                 Some("shims/yarn".into())
+            );
+        }
+
+        #[tokio::test(flavor = "multi_thread")]
+        async fn extracts_auth_token_header() {
+            let sandbox = create_empty_proto_sandbox();
+            sandbox.create_file(".yarnrc.yml", "npmAuthToken: abc123");
+
+            let plugin = sandbox
+                .create_plugin_with_config("yarn-test", |config| {
+                    config.host(HostOS::Linux, HostArch::Arm64);
+                })
+                .await;
+
+            assert_eq!(
+                plugin
+                    .download_prebuilt(DownloadPrebuiltInput {
+                        context: PluginContext {
+                            version: VersionSpec::parse("9.0.0").unwrap(),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    })
+                    .await,
+                DownloadPrebuiltOutput {
+                    archive_prefix: Some("package".into()),
+                    download_url: "https://registry.npmjs.org/npm/-/npm-9.0.0.tgz".into(),
+                    http_headers: FxHashMap::from_iter([(
+                        "Authorization".into(),
+                        "Bearer abc123".into()
+                    )]),
+                    ..Default::default()
+                }
+            );
+        }
+
+        #[tokio::test(flavor = "multi_thread")]
+        async fn extracts_auth_token_header_for_custom_registry() {
+            let sandbox = create_empty_proto_sandbox();
+            sandbox.create_file(
+                ".yarnrc.yml",
+                r#"
+npmAuthToken: xyz789
+
+npmRegistries:
+    //registry.yarnpkg.com:
+        npmAuthToken: abc123
+"#,
+            );
+
+            let plugin = sandbox
+                .create_plugin_with_config("yarn-test", |config| {
+                    config.host(HostOS::Linux, HostArch::Arm64).tool_config(
+                        NodeDepmanPluginConfig {
+                            registry_url: "https://registry.yarnpkg.com".into(),
+                        },
+                    );
+                })
+                .await;
+
+            assert_eq!(
+                plugin
+                    .download_prebuilt(DownloadPrebuiltInput {
+                        context: PluginContext {
+                            version: VersionSpec::parse("9.0.0").unwrap(),
+                            ..Default::default()
+                        },
+                        ..Default::default()
+                    })
+                    .await,
+                DownloadPrebuiltOutput {
+                    archive_prefix: Some("package".into()),
+                    download_url: "https://registry.yarnpkg.com/npm/-/npm-9.0.0.tgz".into(),
+                    http_headers: FxHashMap::from_iter([(
+                        "Authorization".into(),
+                        "Bearer abc123".into()
+                    )]),
+                    ..Default::default()
+                }
             );
         }
     }